// Y-Recorder/src/models/observers.ts
export type Observer = {
  options?: boolean | object

  install(options?: object): void
  uninstall(): void
} & PubSubPattern

export type PubSubPattern = {
  queues: Map<string, Function[]>
  $on(hook: string, action: Function): void
  $off(hook: string, thisAction: Function): void
  $emit(hook: string, ...args): void
}
// ...
export enum DragTypes {
  dragstart = 'dragstart',
  drag = 'drag',
  dragenter = 'dragenter',
  dragover = 'dragover',
  drop = 'drop',
  dragend = 'dragend',
  dragleave = 'dragleave',
}

export enum MouseTypes {
  click = 'click',
  dblclick = 'dblclick',
  move = 'move',
  rightClick = 'rightClick',
  mouseup = 'mouseup',
  mousedown = 'mousedown',
}

export type MouseOptions = {
  click?: boolean
  mousemove?: boolean
}

export type EventOptions = {
  scroll?: boolean
  resize?: boolean
  form?: boolean
  domsWillScroll?: string[]
}

export type Listener = {
  target: Window | Document | ElementX
  event: string
  callback: EventListenerOrEventListenerObject
  options?: AddEventListenerOptions | boolean
}

export type EventReocrd = {
  type: EventTypes
  // scroll
  x?: number
  y?: number
  // resize
  w?: number
  h?: number
  // form change
  target?: number
  tagName?:string
  k?: string
  v?: number | string
}

export type KeyboardReocrd = {
  type: KeyboardTypes
  key?:string
}

export type MouseReocrd = {
  type: MouseTypes|DragTypes
  x?: number
  y?: number
  tag?:string[]
}